# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Rep API",
        "description": "E-Rep Pharmaceutical Platform API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company/events/{eventId}/requests": {
            "get": {
                "tags": [
                    "Company - Events"
                ],
                "summary": "List event registration requests",
                "operationId": "11e1fcfe9b3c42e781de3e5a916befca",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/events/{eventId}/requests/{id}/approve": {
            "post": {
                "tags": [
                    "Company - Events"
                ],
                "summary": "Approve event request",
                "operationId": "d5c0c8ce46d443e10d5db4ecef0995f8",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/events/{eventId}/requests/{id}/reject": {
            "post": {
                "tags": [
                    "Company - Events"
                ],
                "summary": "Reject event request",
                "operationId": "291ad472c9af9a2b8b237e1662494ae3",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/reps": {
            "get": {
                "tags": [
                    "Company - Reps"
                ],
                "summary": "List company medical reps",
                "operationId": "4705fc5a35c679ea8fea96b6b3cb100c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/reps/{id}": {
            "get": {
                "tags": [
                    "Company - Reps"
                ],
                "summary": "Get medical rep",
                "operationId": "9998f56c4beb6d643d05decafe205c48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/reps/{id}/targets": {
            "get": {
                "tags": [
                    "Company - Reps"
                ],
                "summary": "List rep targets with progress",
                "operationId": "855b7dfdaa1a07b980544468be5103c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Rep not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company - Reps"
                ],
                "summary": "Create or update rep target",
                "operationId": "a4f5d8baf1d976bc0a76bdde310404e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "target_type": {
                                        "type": "string",
                                        "enum": [
                                            "meetings",
                                            "samples",
                                            "reviews",
                                            "events",
                                            "doctors"
                                        ]
                                    },
                                    "target_value": {
                                        "type": "integer"
                                    },
                                    "period_start": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "period_end": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Rep not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}